Split out Unix-only tests
